code review clean up

+++ src/main/java/com/sigma/cinematicketpro/service/GenreService.java
@@ -1,9 +1,9 @@
 package com.sigma.cinematicketpro.service;
 
-import com.sigma.cinematicketpro.dto.tmdb.TMDBGenre;
+import com.sigma.cinematicketpro.entity.tmdb.TMDBGenre;
 
 import java.util.Map;
 
 public interface GenreService {
-   Map<Integer, TMDBGenre> getAllGenresMap();
+   Map<Long, TMDBGenre> getAllGenresMap();
 }
